move useCurrentUser from core/hooks to users/hooks (#3831)
