+++ .changeset/happy-paws-join.md
@@ -1,0 +1,5 @@
+---
+"@blitzjs/generator": patch
+---
+
+Move `useCurrentUser` hook from `core/hooks` to `users/hooks` folder
