+++ packages/generator/templates/app/README.md
@@ -90,8 +90,6 @@ __name__
 │   │   ├── components/
 │   │   │   ├── Form.tsx
 │   │   │   └── LabeledTextField.tsx
-│   │   ├── hooks/
-│   │   │   └── useCurrentUser.ts
 │   │   └── layouts/
 │   │       └── Layout.tsx
 │   ├── pages/
@@ -101,6 +99,8 @@ __name__
 │   │   ├── index.test.tsx
 │   │   └── index.tsx
 │   └── users/
+│       ├── hooks/
+│       │   └── useCurrentUser.ts
 │       └── queries/
 │           └── getCurrentUser.ts
 ├── db/
