+++ packages/generator/templates/app/app/users/hooks/useCurrentUser.ts
